requirements.txt
fastapi
numpy
Pillow
tensorflow-cpu
uvicorn
python-multipart
gunicorn
